Add modular `flepimop2` CLI infrastructure
Added the infrastructure for a modular and cohesive `flepimop2` CLI.
Individual commands (i.e. `flepimop2 simulate`) implement the
`flepimop2._cli._cli_command.CliCommand` abstract class which contains
an abstract `run` method for implementation details. The first example
of this implementation is the `SimulateCommand` class, which currently
contains a dummy implementation. Commands that implement this class can
request options/arguments via the `options` attribute which pulls the
details from the `flepimop2._cli._options.COMMAND_OPTIONS` constant.
This results in a cohesive CLI by forcing commands to use the same name
and types for an option/argument. This work is built around `click` and
corresponding unit tests have been added. Also added a `BuildCommand`
for purposes of scaffolding, but only raises a `NotImplementedError`.