Add initial `flepimop2.configuration`
Added an initial outline for the `flepimop2.configuration` module that
will serve as the basis for defining configuration files. Currently the
main elements are:

* The `ConfigurationModel` for defining a complete configuration file.
* The `EngineModel`, `SystemModel`, and `SimulateSpecificationModel` for
  representing the engines, systems, and simulate  sections within a
  `ConfigurationModel`.
* `ParameterSpecificationModel` (currently just an alias for
  `FixedParameterSpecificationModel`), but can be adjusted in the
  future.
* The `FixedParameterSpecificationModel` for representing fixed value
  parameters with convenient short hand for parsed/serialized models.

The current `ConfigurationModel` is very loose and leaves many details
to be filled in. This class subclasses the also added
`YamlSerializableBaseModel` which provides methods for parsing an
instance of this class from a configuration YAML file or serializing a
configuration to a YAML file.

In addition made minor changes:

* Added `pydantic` and `pyyaml` to project dependencies.
* Added `types-PyYAML` and `typing-extensions` to project development
  dependencies.