Integrating logging with CLI commands
Logging is now integrated into the `CliCommand` class. This class now
wraps its `run` method with `__call__` which allows the class to
intercept the `kwargs` given to a CLI command, construct a logger
instance, and log out the `kwargs` given if the verbosity is set high
enough.

Added `log`, `debug`, `info`, `warning`, `error`, and `critical` methods
to `CliCommand` to call the corresponding method of the logger instance.
Also added a `format` method to control the formatting of how `kwargs`
is displayed when the verbosity is high enough. The `format` method is
designed to be easily extended by implementations with something like:

```python
def format(value: Any) -> Any:
    if isinstance(value, bool):
        return "Yes!" if value else "No!"
    return super().format(value)
```