Simplify how `CliCommand`s request opts/args
Changed the `CliCommand.options` abstract property to a class method
that introspects the `run` method to determine the options/arguments to
pull for the CLI command. This works by:

* The private `CliCommand._literal_options` class method which pulls the
  keyword only arguments from the `run` method.
* The public `CliCommand.options` class method calls `_literal_options`
  and optionally adds 'verbosity' if `auto_append_verbosity` is `True`
  and not present in the keyword arguments.
* Finally, after the command is registered and once called the
  `CliCommand.__call__` dunder wraps `CliCommand.run` and determines if
  it wants the verbosity option.

This design allows commands to have a consistent verbosity option by
default that can be utilized via the `self.log/debug/info/...` methods
rather than needing to accept `verbosity` directly. For the few commands
that do not need a verbosity option they can set the
`auto_append_verbosity` class var to `False`.

Also changed the type of `options` from `tuple[str, ...]` to `list[str]`
since it is no longer a class var and it simplifies the internals. This
change alleviates 'noqa' comments that were previously required for
`CliCommand.run`, partially addressing #27.