Normalize `flepimop2` pipeline component APIs
Refactored `flepimop2` implementations of `BackendABC`, `EngineABC`,
`ProcessABC`, and `SystemABC` to also inherit from `ModuleModel` so the
module is both the representation of the configuration and the
implementation. Also adjusted `build` methods to accept only `config:
dict[str, Any] | ModuleModel`. Light reworking of the test suite was
required, but core functionality is still present.

Closes #40, closes #41.