Stateful configuration, context to process action
Modified the `flepimop2` CLI to work with stateful configuration files
by adding a "history" top level key that tracks recent actions. This
allows for successive invocations of a configuration file to share
context. I.e. A user can simulate a model to produce some set of outputs
and then use the out configuration from said invocation to plot with the
process CLI that will have access to the recent simulation outputs.
Stateful configurations is optional, only used if the user explicitly
provides an output for the configuration.

- Added `--out-config` option to simulate/process that will dump the
  configuration file used to the given file for later reuse.
  Incorporated this feature into the external provider integration test.
- Added "history" top level key to the configuration model as a record
  of the actions performed with the configuration file. This is
  represented as a list of `ActionModel`s which now replaces the
  previous `RunMeta` object.
- Configuration is now passed to `ProcessABC.execute` and supporting
  methods to provide context about recent invokations to the process
  action. Namely to connect a process action with recent simulate
  outputs.
- Renamed some variables within simulate/process CLIs for clarity.

Closes #55. Closes #62.