Add external provider package tutorial
Added a development tutorial on how to create an external provider
package. It walks the user through how to setup the package, two
approaches for having their class built, and how to use their package in
a `flepimop2` configuration. The tutorial uses the example of creating a
backend that can handle NumPy's npz file format.

Closes #56.